Fix empty list handling in minidsp custom descriptor #63
Possible fix for https://github.com/3ll3d00d/ezbeq/issues/63 (more details mentioned in the issue).
Seems to work for me. With this BEQ gets applied on Input 1 only of miniDSP 2x4HD with mentioned custom descriptor.
Tested in WSL (do not know how to compile/test on Windows where I really need the fix).

So for empty slots/channels list it should be okay as some miniDSP descriptors internally set that only. But I'm not really sure about usage of groups and if it affects that. Earlier it was getting `None` most of the time but now it will be empty list `[]`.